add robots in a loop function

+++ controllers/kheperaiv_mqp/kheperaiv_mqp.h
@@ -103,7 +103,7 @@ public:
     std::vector<std::vector<std::vector<float>>> path_arr;
     virtual void SetPath(std::vector<std::vector<std::vector<float>>> path_arrki);
 
-    int id;
+    unsigned long id;
 
 
 private:
